[FEAT] Implement new data connectors UI (#1034)
* WIP data connector redesign

* new UI for data connectors complete

* remove old data connector page/cleanup imports

* cleanup of UI and imports

---------
