+++ frontend/src/components/DataConnectorOption/index.jsx
@@ -1,6 +1,3 @@
-import paths from "@/utils/paths";
-import ConnectorImages from "./media";
-
 export default function DataConnectorOption({ slug }) {
   if (!DATA_CONNECTORS.hasOwnProperty(slug)) return null;
   const { path, image, name, description, link } = DATA_CONNECTORS[slug];
@@ -26,22 +23,3 @@ export default function DataConnectorOption({ slug }) {
     </a>
   );
 }
-
-export const DATA_CONNECTORS = {
-  github: {
-    name: "GitHub Repo",
-    path: paths.settings.dataConnectors.github(),
-    image: ConnectorImages.github,
-    description:
-      "Import an entire public or private Github repository in a single click.",
-    link: "https://github.com",
-  },
-  "youtube-transcript": {
-    name: "YouTube Transcript",
-    path: paths.settings.dataConnectors.youtubeTranscript(),
-    image: ConnectorImages.youtube,
-    description:
-      "Import the transcription of an entire YouTube video from a link.",
-    link: "https://youtube.com",
-  },
-};
