+++ frontend/src/components/SettingsSidebar/index.jsx
@@ -15,7 +15,6 @@ import {
   House,
   List,
   FileCode,
-  Plugs,
   Notepad,
   CodeBlock,
   Barcode,
@@ -75,11 +74,10 @@ export default function SettingsSidebar() {
           className={`z-99 fixed top-0 left-0 transition-all duration-500 w-[100vw] h-[100vh]`}
         >
           <div
-            className={`${
-              showBgOverlay
+            className={`${showBgOverlay
                 ? "transition-all opacity-1"
                 : "transition-none opacity-0"
-            }  duration-500 fixed top-0 left-0 ${USER_BACKGROUND_COLOR} bg-opacity-75 w-screen h-screen`}
+              }  duration-500 fixed top-0 left-0 ${USER_BACKGROUND_COLOR} bg-opacity-75 w-screen h-screen`}
             onClick={() => setShowSidebar(false)}
           />
           <div
@@ -192,11 +190,10 @@ const Option = ({
           transition-all duration-[200ms]
           flex flex-grow w-[75%] gap-x-2 py-[6px] px-[12px] rounded-[4px] justify-start items-center
           hover:bg-workspace-item-selected-gradient hover:text-white hover:font-medium
-          ${
-            isActive
+          ${isActive
               ? "bg-menu-item-selected-gradient font-medium border-outline text-white"
               : "hover:bg-menu-item-selected-gradient text-zinc-200"
-          }
+            }
         `}
         >
           {React.cloneElement(icon, { weight: isActive ? "fill" : "regular" })}
@@ -207,9 +204,8 @@ const Option = ({
       </div>
       {!!subOptions && (isActive || hasActiveChild) && (
         <div
-          className={`ml-4 ${
-            hasActiveChild ? "" : "border-l-2 border-slate-400"
-          } rounded-r-lg`}
+          className={`ml-4 ${hasActiveChild ? "" : "border-l-2 border-slate-400"
+            } rounded-r-lg`}
         >
           {subOptions}
         </div>
@@ -303,14 +299,6 @@ const SidebarOptions = ({ user = null }) => (
       user={user}
       flex={true}
       allowedRole={["admin"]}
-    />
-    <Option
-      href={paths.settings.dataConnectors.list()}
-      btnText="Data Connectors"
-      icon={<Plugs className="h-5 w-5 flex-shrink-0" />}
-      user={user}
-      flex={true}
-      allowedRole={["admin", "manager"]}
     />
     <Option
       href={paths.settings.embedSetup()}
