+++ frontend/src/utils/paths.js
@@ -125,16 +125,5 @@ export default {
     embedChats: () => {
       return `/settings/embed-chats`;
     },
-    dataConnectors: {
-      list: () => {
-        return "/settings/data-connectors";
-      },
-      github: () => {
-        return "/settings/data-connectors/github";
-      },
-      youtubeTranscript: () => {
-        return "/settings/data-connectors/youtube-transcript";
-      },
-    },
   },
 };
